feature: extract image card and status badge
