+++ src/meal-ui/src/components/Card/index.js
@@ -1,1 +1,4 @@
+import ImageCard from './ImageCard';
+
 export { default } from './Card';
+export { ImageCard };
